pyproject.toml
[project]
name = "goldenzona-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.10.0",
    "boto3>=1.40.74",
    "botocore>=1.40.74",
    "celery>=5.5.3",
    "cryptography>=46.0.3",
    "defusedxml>=0.7.1",
    "django>=5.2.8",
    "django-allauth>=65.13.0",
    "django-cities-light>=3.10.2",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-schema-graph>=3.1.0",
    "django-smtp-ssl>=1.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-yasg>=1.21.11",
    "eth-account>=0.13.7",
    "eth-hash>=0.7.1",
    "gunicorn>=23.0.0",
    "idna>=3.11",
    "jmespath>=1.0.1",
    "mnemonic>=0.21",
    "oauth2-provider>=0.0",
    "oauthlib>=3.3.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "s3transfer>=0.14.0",
    "sqlparse>=0.5.3",
    "tzdata>=2025.2",
    "uvicorn>=0.38.0",
    "web3>=7.14.0",
    "whitenoise>=6.11.0",
]
